09-agent-eval/pyproject.toml
[project]
name = "agent-eval"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.29.5",
    "llama-index-core>=0.12.40",
    "llama-index-llms-openai>=0.4.2",
    "python-dotenv>=1.1.0",
    "ragas>=0.2.15",
]
